[transitions] ActionConfigs to transitions
