--- src/main/java/com/community/tools/util/statemachie/actions/questions/FirstQuestionAction.java
@@ -1,23 +1,0 @@
-package com.community.tools.util.statemachie.actions.questions;
-
-import com.community.tools.service.slack.SlackService;
-import com.community.tools.util.statemachie.Event;
-import com.community.tools.util.statemachie.State;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.statemachine.StateContext;
-import org.springframework.statemachine.action.Action;
-
-public class FirstQuestionAction implements Action<State, Event> {
-
-  @Value("${firstQuestion}")
-  private String firstQuestion;
-  @Autowired
-  private SlackService slackService;
-
-  @Override
-  public void execute(StateContext<State, Event> stateContext) {
-    String user = stateContext.getExtendedState().getVariables().get("id").toString();
-    slackService.sendBlocksMessage(slackService.getUserById(user), firstQuestion);
-  }
-}
