+++ src/main/java/com/community/tools/util/statemachie/actions/transitions/configs/tasks/GetTheFirstTaskActionTransition.java
@@ -1,0 +1,46 @@
+package com.community.tools.util.statemachie.actions.transitions.configs.tasks;
+
+import static com.community.tools.util.statemachie.Event.GET_THE_FIRST_TASK;
+import static com.community.tools.util.statemachie.State.ADDED_GIT;
+import static com.community.tools.util.statemachie.State.GOT_THE_FIRST_TASK;
+
+import com.community.tools.service.slack.SlackService;
+import com.community.tools.util.statemachie.Event;
+import com.community.tools.util.statemachie.State;
+import com.community.tools.util.statemachie.actions.Transition;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.statemachine.StateContext;
+import org.springframework.statemachine.action.Action;
+import org.springframework.statemachine.annotation.WithStateMachine;
+import org.springframework.statemachine.config.builders.StateMachineTransitionConfigurer;
+
+@WithStateMachine
+public class GetTheFirstTaskActionTransition implements Transition {
+
+  @Value("${getFirstTask}")
+  private String getFirstTask;
+
+  @Autowired
+  private SlackService slackService;
+
+  @Autowired
+  private Action<State, Event> errorAction;
+
+  @Override
+  public void configure(
+      StateMachineTransitionConfigurer<State, Event> transitions) throws Exception {
+    transitions
+        .withExternal()
+        .source(ADDED_GIT)
+        .target(GOT_THE_FIRST_TASK)
+        .event(GET_THE_FIRST_TASK)
+        .action(this, errorAction);
+  }
+
+  @Override
+  public void execute(StateContext<State, Event> stateContext) {
+    String user = stateContext.getExtendedState().getVariables().get("id").toString();
+    slackService.sendBlocksMessage(slackService.getUserById(user), getFirstTask);
+  }
+}
