--- src/main/java/com/community/tools/util/statemachie/actions/verifications/AgreeLicenseAction.java
@@ -1,24 +1,0 @@
-package com.community.tools.util.statemachie.actions.verifications;
-
-import com.community.tools.service.slack.SlackService;
-import com.community.tools.util.statemachie.Event;
-import com.community.tools.util.statemachie.State;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.statemachine.StateContext;
-import org.springframework.statemachine.action.Action;
-
-
-public class AgreeLicenseAction implements Action<State, Event> {
-
-  @Value("${addGitName}")
-  private String addGitName;
-  @Autowired
-  private SlackService slackService;
-
-  @Override
-  public void execute(StateContext<State, Event> stateContext) {
-    String user = stateContext.getExtendedState().getVariables().get("id").toString();
-    slackService.sendBlocksMessage(slackService.getUserById(user), addGitName);
-  }
-}
