--- src/main/java/com/community/tools/util/statemachie/actions/verifications/VerificationLoginAction.java
@@ -1,41 +1,0 @@
-package com.community.tools.util.statemachie.actions.verifications;
-
-import com.community.tools.service.github.GitHubConnectService;
-import com.community.tools.service.github.GitHubService;
-import com.community.tools.service.slack.SlackService;
-import com.community.tools.util.statemachie.Event;
-import com.community.tools.util.statemachie.State;
-
-import java.io.IOException;
-
-import org.kohsuke.github.GHUser;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.statemachine.StateContext;
-import org.springframework.statemachine.action.Action;
-
-public class VerificationLoginAction implements Action<State, Event> {
-
-  @Value("${askAboutProfile}")
-  private String askAboutProfile;
-  @Autowired
-  private SlackService slackService;
-  @Autowired
-  private GitHubConnectService gitHubConnectService;
-  @Autowired
-  private GitHubService gitHubService;
-
-  @Override
-  public void execute(StateContext<State, Event> context) {
-    String user = context.getExtendedState().getVariables().get("id").toString();
-    String nickname = context.getExtendedState().getVariables().get("gitNick").toString();
-    try {
-      GHUser userGitLogin = gitHubService.getUserByLoginInGitHub(nickname);
-      slackService.sendPrivateMessage(slackService.getUserById(user),
-              askAboutProfile + "\n" + userGitLogin.getHtmlUrl().toString());
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-
-  }
-}
